requirements.txt
torch>=2.0.0
torchvision>=0.15.0

numpy>=1.23

Pillow>=9.4

matplotlib>=3.7